requirements.txt
twitchio==2.6.0
pyautogui==0.9.54
pyperclip==1.8.2
python-dotenv==0.20.0
python-win32==306
colorama==0.4.6
keyboard==0.13.5